+++ cacheai/__init__.py
@@ -1,4 +1,0 @@
-from .core import CacheAI
-
-__all__ = ["CacheAI"]
-__version__ = "0.1.0"